@@ -21,11 +21,11 @@
     ],
 	"config": {
         "platform": {
-            "php": "5.6.32"
+            "php": "7.0"
         }
     },
     "require": {
-        "php": ">=5.5",
+        "php": ">=7.0",
         "ext-curl": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
@@ -34,14 +34,14 @@
     "require-dev": {
 	    "aspose/storage-sdk-php": "18.*",
         "squizlabs/php_codesniffer": "3.*",
-        "phpunit/phpunit": "^4.8",
+        "phpunit/phpunit": "6.0",
 		"behat/behat": "3.4.*",
 		"friendsofphp/php-cs-fixer": "2.10.*"
     },
 	"autoload": {
         "psr-0": {
             "GroupDocs\\Assembly\\": "Assembly/",
-            "BaseTest": "tests/"
+            "BaseTest\\": "tests/"
         }
     }
 }
